[DEV-2.3] Change mode by pressing the button on UI
## Describe Your Change

In this user story, I have created a button on UI for user in order to change between light mode and dark mode if they want. To implement the goal, I have made some change on "views".

# IMode
The IMode strategy interface contains a method setMode() for lightMode() class and darkMode() class to implement different algorithms for TetrisView.

# lightMode() and darkMode()
The classes to implement different algorithms.

# Mode()
This class is to communicate via the IMode strategy interface; it define the interface of interest to client.

# TetrisView
I have add an attribute "curr" to determine the current mode.
I have add a modeButton in the control HBox.
When the modeButton is pressed, the attribute "curr" will be check and change to the match mode. The mode will change on background, canvas, buttons and font.

## User story Number
2.3 Change Mode

## Checklist before requesting a review

[√] I have performed a self-view of my code.
[√] If it’s a core feature, I have added thorough tests.